requirements.txt
requests
pandas
numpy
matplotlib
seaborn
biopython
jupyter
pytest
